+++ client-interface/components/admin/certificates/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './useAIEvaluationProgress';
-export * from './useRecipientSelection';
+export * from './useCertificateHooks';
