+++ server/src/index.js
@@ -112,9 +112,6 @@ async function start() {
     if (process.env.CERTIFICATE_WORKER_DISABLED !== 'true') {
       require('./workers/certificateWorker').start();
     }
-    if (process.env.AI_EVAL_WORKER_DISABLED !== 'true') {
-      require('./workers/aiEvaluationWorker').start();
-    }
     if (process.env.NOTIFICATION_SCHEDULER_DISABLED !== 'true') {
       notificationScheduler.start();
     }
